Relocate the DataSourceError to make it can be used cross plugin folder
